backend/requirements.txt
asgiref==3.8.1
django==5.2.3
djangorestframework==3.16.0
sqlparse==0.5.3
typing-extensions==4.14.0
django-cors-headers==4.3.1
pvlib==0.10.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
